feat: ios app builds
