+++ app.json
@@ -2,7 +2,7 @@
   "displayName": "Learner Credential Wallet",
   "expo": {
     "name": "Learner Credential Wallet",
-    "slug": "Learner Credential Wallet",
+    "slug": "learner-credential-wallet",
     "version": "1.0.0",
     "orientation": "portrait",
     "icon": "./app/assets/icon.png",
@@ -20,7 +20,13 @@
     ],
     "ios": {
       "supportsTablet": true,
-      "bundleIdentifier": "edu.wallet"
+      "bundleIdentifier": "edu.mit.eduwallet",
+      "entitlements": {
+        "com.apple.security.application-groups": [
+          "group.edu.mit.eduwallet"
+        ]
+      },
+      "associatedDomains": ["applinks:lcw.app/mobile"]
     },
     "android": {
       "adaptiveIcon": {
@@ -31,6 +37,16 @@
     },
     "web": {
       "favicon": "./app/assets/favicon.png"
-    }
+    },
+    "plugins": [
+      [
+        "react-native-vision-camera",
+        {
+          "cameraPermissionText": "$(PRODUCT_NAME) needs access to your Camera to scan QR codes.",
+          "enableMicrophonePermission": false,
+          "enableCodeScanner": true
+        }
+      ]
+    ]
   }
 }
