+++ app/components/VerificationCard/VerificationCard.tsx
@@ -4,10 +4,10 @@ import { View, Text, TouchableOpacity } from 'react-native';
 import { MaterialCommunityIcons, MaterialIcons } from '@expo/vector-icons';
 
 import { useDynamicStyles, useVerifyCredential } from '../../hooks';
-import { VerifyPayload } from '../../lib/verifiableObject';
 import dynamicStyleSheet from './VerificationCard.styles';
 import { navigationRef } from '../../navigation';
 import { CredentialRecordRaw } from '../../model';
+import {VerifyPayload} from '../../lib/verificationResultFor';
 
 type CommonProps = {
   isButton?: boolean;
