+++ app/navigation/RootNavigation/RootNavigation.d.tsx
@@ -1,11 +1,11 @@
 import type { NavigatorScreenParams } from '@react-navigation/native';
 import type { StackScreenProps } from '@react-navigation/stack';
 import type { CredentialRecordRaw } from '../../model/credential';
-import type { VerifyPayload } from '../../lib/verifiableObject';
 import type { HomeNavigationParamList, AcceptCredentialsNavigationParamList, ExchangeCredentialsNavigationParamList } from '..';
 import { Credential } from '../../types/credential';
 import { ProfileRecordRaw } from '../../model';
 import { CredentialSelectionScreenParams, ProfileSelectionScreenParams, QRScreenParams, ViewSourceScreenParams } from '../../screens';
+import {VerifyPayload} from '../../lib/verificationResultFor';
 
 export type RootNavigationParamsList = {
   HomeNavigation: NavigatorScreenParams<HomeNavigationParamList>;
