+++ app/screens/QRScreen/QRScreen.tsx
@@ -1,8 +1,6 @@
-import React, { useState, useRef } from 'react';
-import { Alert, Text, Linking } from 'react-native';
-import { View, useWindowDimensions } from 'react-native';
-import QRCodeScanner from 'react-native-qrcode-scanner';
-import { BarCodeReadEvent, RNCameraProps } from 'react-native-camera';
+import React, { useState,  } from 'react';
+import {  Text, Linking } from 'react-native';
+import { View, useWindowDimensions, StyleSheet } from 'react-native';
 
 import { ConfirmModal } from '../../components';
 import { NavHeader } from '../../components';
@@ -10,14 +8,17 @@ import { QRScreenProps } from './QRScreen.d';
 import dynamicStyleSheet from './QRScreen.styles';
 import { errorMessageFrom } from '../../lib/error';
 import { useDynamicStyles } from '../../hooks';
-
-export default function QRScreen({ navigation, route }: QRScreenProps): JSX.Element {
+import {Camera, CodeType, useCameraDevice, useCameraFormat, useCodeScanner} from 'react-native-vision-camera';
+import {useIsFocused} from '@react-navigation/native';
+import {useAppState} from '@react-native-community/hooks';
+import {Code} from 'react-native-vision-camera/src/CodeScanner';
+const codeTypes: CodeType[] = ['qr'];
+export default function QRScreen({ navigation, route }: QRScreenProps)  {
   const { styles } = useDynamicStyles(dynamicStyleSheet);
   const { onReadQRCode, instructionText } = route.params;
 
   const [errorMessage, setErrorMessage] = useState('');
   const { width } = useWindowDimensions();
-  const scannerRef = useRef<QRCodeScanner>(null);
 
   const errorModalOpen = errorMessage !== '';
 
@@ -29,9 +30,9 @@ export default function QRScreen({ navigation, route }: QRScreenProps): JSX.Elem
     );
   }
 
-  async function onRead({ data: text }: BarCodeReadEvent) {
+  async function onRead(codes: Code[]) {
     try {
-      await onReadQRCode(text);
+      // await onReadQRCode(text);
     } catch (err) {
       setErrorMessage(errorMessageFrom(err));
     }
@@ -40,7 +41,7 @@ export default function QRScreen({ navigation, route }: QRScreenProps): JSX.Elem
 
   function onRequestModalClose() {
     setErrorMessage('');
-    setTimeout(() => scannerRef.current?.reactivate(), 1000);
+    // setTimeout(() => scannerRef.current?.reactivate(), 1000);
   }
 
   function goToSettings() {
@@ -51,50 +52,84 @@ export default function QRScreen({ navigation, route }: QRScreenProps): JSX.Elem
     navigation.goBack();
   }
 
-  function NoCameraPermission(): JSX.Element {
-    return (
-      <View>
-        { 
-          Alert.alert(
-            'Camera Access is Off', 
-            'Please go into your camera settings and enable access for this app',
-            [
-              {
-                text: 'Go Back',
-                onPress: navGoBack
-              },
-              {
-                text: 'Settings',
-                onPress: goToSettings
-              }
-            ])
-        }
-      </View>
-    );
+  // function NoCameraPermission(): JSX.Element {
+  //   return (
+  //     <View>
+  //       {
+  //         Alert.alert(
+  //           'Camera Access is Off',
+  //           'Please go into your camera settings and enable access for this app',
+  //           [
+  //             {
+  //               text: 'Go Back',
+  //               onPress: navGoBack
+  //             },
+  //             {
+  //               text: 'Settings',
+  //               onPress: goToSettings
+  //             }
+  //           ])
+  //       }
+  //     </View>
+  //   );
+  // }
+
+  const codeScanner = useCodeScanner({
+    codeTypes: codeTypes,
+    onCodeScanned: onRead,
+  });
+
+  const device = useCameraDevice('back');
+
+  const format = useCameraFormat(device, [
+    {photoResolution: {width: 1920, height: 1080}},
+    {videoResolution: {width: 1920, height: 1080}},
+  ]);
+  const isFocused = useIsFocused();
+  const appState = useAppState();
+  const isActive = isFocused && appState === 'active';
+
+  if (!device) {
+    return null;
+    // TODO should return <NoCameraView />;
   }
+
 
   return (
     <View style={styles.scannerBody}>
       <NavHeader title="QR Code Scanner" goBack={navigation.goBack} />
-      <QRCodeScanner
-        ref={scannerRef}
-        onRead={onRead}
-        topContent={<Instructions />}
-        topViewStyle={styles.instructionContainer}
-        bottomViewStyle={styles.emptyContainer}
-        cameraStyle={styles.cameraStyle}
-        markerStyle={[styles.markerStyle, {
-          width: width * 0.8,
-          height: width * 0.8,
-        }]}
-        notAuthorizedView={<NoCameraPermission />}
-        cameraProps={{
-          accessibilityLabel: 'QR Code Scanner, Camera Active',
-          accessible: true,
-          notAuthorizedView: <NoCameraPermission />
-        } as RNCameraProps}
-        showMarker
+      <Camera
+        isActive={isActive}
+        device={device}
+        format={format}
+        codeScanner={codeScanner}
+        style={StyleSheet.absoluteFill}
+        // onError={logger.error}
+        photoHdr={false}
+        orientation="portrait"
+        photo
+        video={false}
+        audio={false}
       />
+      {/*<QRCodeScanner*/}
+      {/*  ref={scannerRef}*/}
+      {/*  onRead={onRead}*/}
+      {/*  topContent={<Instructions />}*/}
+      {/*  topViewStyle={styles.instructionContainer}*/}
+      {/*  bottomViewStyle={styles.emptyContainer}*/}
+      {/*  cameraStyle={styles.cameraStyle}*/}
+      {/*  markerStyle={[styles.markerStyle, {*/}
+      {/*    width: width * 0.8,*/}
+      {/*    height: width * 0.8,*/}
+      {/*  }]}*/}
+      {/*  notAuthorizedView={<NoCameraPermission />}*/}
+      {/*  cameraProps={{*/}
+      {/*    accessibilityLabel: 'QR Code Scanner, Camera Active',*/}
+      {/*    accessible: true,*/}
+      {/*    notAuthorizedView: <NoCameraPermission />*/}
+      {/*  } as RNCameraProps}*/}
+      {/*  showMarker*/}
+      {/*/>*/}
       <ConfirmModal
         open={errorModalOpen}
         onRequestClose={onRequestModalClose}
