+++ app/store/index.ts
@@ -1,10 +1,10 @@
-import { configureStore, createAsyncThunk } from '@reduxjs/toolkit';
+import { configureStore } from '@reduxjs/toolkit';
 
 import wallet from './slices/wallet';
-import did, { _getAllDidRecords } from './slices/did';
+import did  from './slices/did';
 import credentialFoyer from './slices/credentialFoyer';
-import credential, { _getAllCredentialRecords } from './slices/credential';
-import profile, { _getAllProfileRecords } from './slices/profile';
+import credential  from './slices/credential';
+import profile  from './slices/profile';
 
 const store = configureStore({
   reducer: {
@@ -19,12 +19,7 @@ const store = configureStore({
   }),
 });
 
-export const getAllRecords = createAsyncThunk('getAllRecords', async (_, { dispatch }) => {
-  await dispatch(_getAllDidRecords());
-  await dispatch(_getAllCredentialRecords());
-  await dispatch(_getAllProfileRecords());
-});
-
+export {getAllRecords} from './getAllRecords';
 export default store;
 
 export type RootState = ReturnType<typeof store.getState>
