+++ app/styles/theme/index.ts
@@ -4,9 +4,10 @@ import { lightTheme } from './themes/lightTheme';
 export const defaultTheme = darkTheme;
 
 export const themes = {
-  darkTheme, 
+  darkTheme,
   lightTheme,
 };
+import { ThemeType } from './index.d';
 
 export type { ThemeType } from './index.d';
 
@@ -15,3 +16,10 @@ export * from './themeContext';
 
 /* Temporary static definition */
 export default defaultTheme;
+export function findThemeBy(themeName: string | null): ThemeType | null {
+  if (themeName === null) {
+    return null;
+  }
+
+  return Object.values(themes).find(({ name }) => name === themeName) || null;
+}
