+++ app/styles/theme/themeName.ts
@@ -1,6 +1,5 @@
 import AsyncStorage from '@react-native-async-storage/async-storage';
 
-import { themes, ThemeType } from '.';
 import { AsyncStorageKey } from '../../lib/asyncStorage';
 
 export async function loadThemeName(): Promise<string | null> {
@@ -11,10 +10,3 @@ export async function saveThemeName(themeName: string): Promise<void> {
   return AsyncStorage.setItem(AsyncStorageKey.ThemeName, themeName);
 }
 
-export function findThemeBy(themeName: string | null): ThemeType | null {
-  if (themeName === null) {
-    return null;
-  }
-
-  return Object.values(themes).find(({ name }) => name === themeName) || null;
-}
