+++ index.js
@@ -1,4 +1,5 @@
 import './shim.js';
+import 'react-native-url-polyfill/auto';
 import 'react-native-gesture-handler';
 import 'react-native-get-random-values';
 import { registerRootComponent } from 'expo';
