+++ metro.config.js
@@ -3,6 +3,6 @@ const { getDefaultConfig } = require('expo/metro-config');
 
 const defaultConfig = getDefaultConfig(__dirname);
 
-defaultConfig.resolver.assetExts.push("cjs");
+defaultConfig.resolver.assetExts.push('cjs');
 
 module.exports = defaultConfig;
