+++ package.json
@@ -3,24 +3,42 @@
   "name": "learner-credential-wallet",
   "version": "1.0.0",
   "scripts": {
-    "start": "react-native start",
+    "start": "npx expo start --dev-client -c",
     "android": "react-native run-android",
     "ios": "react-native run-ios",
     "lint": "eslint ./app ./App.tsx && npm run compile",
     "test": "jest",
     "compile": "tsc --noEmit -p tsconfig.json && echo ✓ Passed lint check",
     "postinstall": "npx patch-package",
-    "fix-deps": "npx expo install --check"
+    "fix-deps": "npx expo install --check",
+    "prebuild": "EXPO_NO_GIT_STATUS=1 npx expo prebuild --clean -p ios"
+  },
+  "expo": {
+    "autolinking": {
+      "android": {
+        "exclude": [
+          "@unimodules/core",
+          "@unimodules/react-native-adapter"
+        ]
+      },
+      "ios": {
+        "exclude": [
+          "@unimodules/core",
+          "@unimodules/react-native-adapter"
+        ]
+      }
+    }
   },
   "dependencies": {
     "@digitalcredentials/cborld": "^4.3.4",
-    "@digitalcredentials/lru-memoize": "^2.1.4",
     "@digitalcredentials/crypto-ld": "^7.0.3",
     "@digitalcredentials/did-io": "^1.0.2",
     "@digitalcredentials/did-method-key": "^2.0.3",
-    "@digitalcredentials/ed25519-signature-2020": "^3.0.1",
+    "@digitalcredentials/ed25519-signature-2020": "^3.0.2",
     "@digitalcredentials/ed25519-verification-key-2020": "^4.0.0",
     "@digitalcredentials/issuer-registry-client": "^2.0.0",
+    "@digitalcredentials/jsonld-signatures": "^9.3.2",
+    "@digitalcredentials/lru-memoize": "^2.1.4",
     "@digitalcredentials/minimal-cipher": "^5.1.2",
     "@digitalcredentials/security-document-loader": "^3.1.0",
     "@digitalcredentials/vc": "^6.0.0",
@@ -30,9 +48,10 @@
     "@expo-google-fonts/roboto-mono": "^0.2.2",
     "@expo-google-fonts/rubik": "^0.2.0",
     "@expo/vector-icons": "^14.0.0",
-    "@interop/did-web-resolver": "^2.2.1",
+    "@interop/did-web-resolver": "^3.0.1",
     "@react-native-async-storage/async-storage": "^1.17.7",
     "@react-native-clipboard/clipboard": "^1.10.0",
+    "@react-native-community/hooks": "^3.0.0",
     "@react-native-community/masked-view": "0.1.10",
     "@react-native-community/netinfo": "11.1.0",
     "@react-navigation/bottom-tabs": "^6.0.7",
@@ -41,16 +60,18 @@
     "@reduxjs/toolkit": "^1.6.1",
     "big-integer": "^1.6.49",
     "bson": "4.4.1",
-    "expo": "^50.0.0-preview.11",
+    "expo": "~50.0.2",
+    "expo-dev-client": "~3.3.6",
     "expo-font": "~11.10.1",
     "expo-secure-store": "~12.8.1",
-    "expo-splash-screen": "~0.26.2",
+    "expo-splash-screen": "~0.26.3",
     "expo-status-bar": "~1.11.1",
-    "expo-updates": "~0.24.7",
+    "expo-updates": "~0.24.8",
     "handlebars": "^4.7.8",
     "html-entities": "^2.3.2",
     "json-canonicalize": "^1.0.4",
     "jsonpath-plus": "^7.2.0",
+    "ky": "^1.2.0",
     "moment": "^2.29.1",
     "query-string": "^7.1.0",
     "react": "18.2.0",
@@ -59,14 +80,14 @@
     "react-native": "0.73.2",
     "react-native-circular-progress": "^1.3.7",
     "react-native-device-info": "^10.12.0",
-    "react-native-document-picker": "^9.1.0",
+    "react-native-document-picker": "^8.2.1",
     "react-native-dropdown-picker": "^5.4.2",
     "react-native-elements": "^3.4.2",
     "react-native-exit-app": "^2.0.0",
     "react-native-file-logger": "^0.4.0",
     "react-native-fs": "^2.20.0",
     "react-native-gesture-handler": "~2.14.0",
-    "react-native-get-random-values": "~1.8.0",
+    "react-native-get-random-values": "~1.10.0",
     "react-native-html-to-pdf": "^0.12.0",
     "react-native-keyboard-aware-scroll-view": "^0.9.4",
     "react-native-keychain": "^8.1.1",
@@ -75,6 +96,8 @@
     "react-native-pager-view": "6.2.3",
     "react-native-paper": "^4.12.3",
     "react-native-qrcode-svg": "^6.1.2",
+    "react-native-quick-crypto": "^0.6.1",
+    "react-native-reanimated": "~3.6.0",
     "react-native-receive-sharing-intent": "^2.0.0",
     "react-native-safe-area-context": "4.8.2",
     "react-native-screens": "~3.29.0",
@@ -83,17 +106,18 @@
     "react-native-storage": "^1.0.1",
     "react-native-svg": "14.1.0",
     "react-native-swipeable": "^0.6.0",
+    "react-native-url-polyfill": "^2.0.0",
     "react-native-uuid": "^2.0.1",
-    "react-native-vision-camera": "^3.8.1",
+    "react-native-vision-camera": "^3.8.2",
     "react-redux": "^7.2.4",
-    "realm": "^10.6.1",
+    "realm": "^11.10.2",
     "rn-animated-ellipsis": "^2.1.3",
     "text-encoding": "^0.7.0",
     "util": "^0.10.4",
     "validator": "^13.11.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.9.0",
+    "@react-native/eslint-config": "^0.73.2",
     "@types/crypto-js": "^4.1.1",
     "@types/jest": "^29.5.5",
     "@types/react": "~18.2.45",
@@ -103,7 +127,6 @@
     "@typescript-eslint/parser": "^6.16.0",
     "babel-plugin-module-resolver": "^5.0.0",
     "eslint": "^8.56.0",
-    "@react-native/eslint-config": "^0.73.2",
     "eslint-plugin-react": "^7.24.0",
     "eslint-plugin-react-native": "^3.11.0",
     "jest": "^29.6.3",
